feat(cloud-service): create all data option to excel download
